Data Journalism and D3
D3 Dabbler

Create a scatter plot between two of the data variables such as Healthcare vs. Poverty or Smokers vs. Age.

Create a scatter plot that represents each state with circle elements. The scatter plot should ultimately appear like the image at the top of this section.

Include state abbreviations in the circles.

Create and situate the axes and labels to the left and bottom of the chart.

Impress the Boss

1. More Data, More Dynamics
You're going to include more demographics and more risk factors. Place additional labels in your scatter plot and give them click events so that your users can decide which data to display. Animate the transitions for your circles' locations as well as the range of your axes. Do this for two risk factors for each axis. Or, for an extreme challenge, create three for each axis.

2. Incorporate d3-tip
While the ticks on the axes allow us to infer approximate values for each circle, it's impossible to determine the true value without adding another layer of data. Enter tooltips: developers can implement these in their D3 graphics to reveal a specific element's data when the user hovers their cursor over the element. Add tooltips to your circles and display each tooltip with the data that the user has selected.